Moving editing form to separate component to allow editing more than one post at a time
