+++ src/components/postBox/PostUpdate.js
@@ -1,16 +1,12 @@
 import React from "react";
-import Swal from "sweetalert2";
-import { useContext, useEffect, useState } from "react";
-import { getAllFeeds } from "../../functions/getAllFeeds";
-import { AuthenticationContext } from "../../contexts/AuthenticationContext";
+import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import ButtonIcon from "../buttons/ButtonIcon";
 import { MdEdit, MdCancel } from "react-icons/md";
+import EditPostForm from "./EditPostForm";
 
 export default function PostUpdate({ setFeeds, post }) {
-  const { setIsLoggedIn } = useContext(AuthenticationContext);
   const [loggedInUser, setLoggedInUser] = useState(null);
-  const [postContent, setPostContent] = useState(post.postText);
   const { username } = useParams();
 
   const [showEdit, setShowEdit] = useState(false);
@@ -23,57 +19,19 @@ export default function PostUpdate({ setFeeds, post }) {
     }
   }, []);
 
-  async function updatePost(event, postID) {
-    event.preventDefault();
-
-    try {
-      const response = await fetch(`http://localhost:5050/posts`, {
-        method: "PATCH",
-        headers: {
-          "content-type": "application/json",
-        },
-        credentials: "include",
-        body: JSON.stringify({ postText: postContent, id: postID }),
-      });
-
-      if (response.status === 401) {
-        await Swal.fire({
-          icon: "error",
-          text: "Please login to be able to edit posts!",
-        });
-        localStorage.removeItem("loggedInUser");
-        setIsLoggedIn(false);
-        return;
-      }
-
-      if (response.status === 200) {
-        const { feedList } = await getAllFeeds(username);
-        setFeeds(feedList);
-        setShowEdit(false);
-        return;
-      }
-    } catch (error) {
-      console.error(error);
-      await Swal.fire({
-        icon: "error",
-        text: "Something went wrong, failed to connect to server!",
-      });
-      setIsLoggedIn("serverError");
-      return;
-    }
-  }
-
   return (
-    <div className="postUpdate">
+    <div className="postBox_description">
       {username !== loggedInUser ? null : (
-        <div className="postBox_buttons">
+        <div className="postBox_description_postEdit">
           {showEdit ? (
             <ButtonIcon
+              className={"editPost cross"}
               handleClick={(event) => setShowEdit(false)}
               icon={<MdCancel />}
             />
           ) : (
             <ButtonIcon
+              className={"editPost"}
               handleClick={(event) => setShowEdit(true)}
               icon={<MdEdit />}
             />
@@ -81,21 +39,14 @@ export default function PostUpdate({ setFeeds, post }) {
         </div>
       )}
       {showEdit ? (
-        <form>
-          <textarea
-            className="addCommentForm_textarea"
-            onChange={(e) => setPostContent(e.target.value)}
-            value={postContent}
-          ></textarea>
-          <button
-            className="addCommentForm_button"
-            onClick={(event) => updatePost(event, post._id)}
-          >
-            Edit Post
-          </button>
-        </form>
+        <EditPostForm
+          post={post}
+          username={username}
+          setFeeds={setFeeds}
+          setShowEdit={setShowEdit}
+        />
       ) : (
-        <p className="postBox_description">{post.postText}</p>
+        <p className="postBox_description_text">{post.postText}</p>
       )}
     </div>
   );
